requirements.txt
alembic==1.13.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.3.3
MarkupSafe==2.0.1
Pillow==8.3.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
twilio==6.63.0
typing_extensions==4.11.0
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
